+++ src/assets/styles/_buttons.scss
@@ -23,10 +23,6 @@
     background: $accent-primary;
     border-radius: 7px;
     color: white;
-
-    &:hover {
-        background-color: $accent;
-    }
 }
 
 .btn-tertiary {
@@ -37,8 +33,8 @@
 
 .btn-secondary, .btn-secondary:hover{
     background-color: white;
-    color: $accent;
-    border: $accent solid 2px;
+    color: $accent-primary;
+    border: $accent-primary solid 2px;
 }
 
 .btn-upload {
