+++ src/assets/styles/_variables.scss
@@ -14,8 +14,8 @@ $accent-secondary: #2FF596;
 
 $danger: #EE2C43;
 
-$font-family-sans-serif: "Nunito, Helvetica Neue, Tahoma, Arial, San-Serif";
-$font-family-serif:      "Nunito, Helvetica Neue, Tahoma, Arial, San-Serif";
-$font-family-monospace:  "Nunito, Helvetica Neue, Tahoma, Arial, San-Serif";
-$font-family-base:       "Nunito, Helvetica Neue, Tahoma, Arial, San-Serif";
+$font-family-sans-serif: "Poppins, Helvetica Neue, Tahoma, Arial, San-Serif";
+$font-family-serif:      "Poppins, Helvetica Neue, Tahoma, Arial, San-Serif";
+$font-family-monospace:  "Poppins, Helvetica Neue, Tahoma, Arial, San-Serif";
+$font-family-base:       "Poppins, Helvetica Neue, Tahoma, Arial, San-Serif";
