+++ src/assets/styles/styles.scss
@@ -3,4 +3,5 @@
 @import "./_reset.scss";
 @import "./_typography.scss";
 @import "./_variables.scss";
+@import "./_form.scss";
