--- src/pages/addstudent/addstudent.scss
@@ -1,7 +1,0 @@
-@import "../../assets/styles/variables";
-
-.add-student{
-    background-color: $primary;
-    width: 80%;
-    float: right ;
-}
