+++ src/pages/students/student-add/addstudent.scss
@@ -1,0 +1,5 @@
+@import "../../../assets/styles/variables";
+
+.add-student{
+    background-color: $color-primary;
+}
