+++ src/pages/students/student-add/addstudenttopnav/addstudenttopnav.scss
@@ -1,5 +1,5 @@
-@import "../../../assets/styles/nav.scss";
-@import "../../../assets/styles/typography";
+@import "../../../../assets/styles/nav.scss";
+@import "../../../../assets/styles/typography";
 
 .topnav{
     height: 120px;
@@ -7,8 +7,8 @@
     font: $font-primary;
 
     &__title{
-    font: $font-primary;
-    font-weight: 700;
+        font: $font-primary;
+        font-weight: 700;
     }
 
     &__button{
