+++ src/pages/students/students-list/pagenavigator/pagenavigator.jsx
@@ -1,9 +1,10 @@
 import React from 'react';
-import LeftArrow from "../../../assets/images/left.png";
-import RightArrow from "../../../assets/images/right.png";
+import LeftArrow from "../../../../assets/images/left.png";
+import RightArrow from "../../../../assets/images/right.png";
 import "./pagenavigator.scss";
 
 const PageNavigator = (props) => {
+    
     const {totalNumberStudents, switchToPreviousPage, switchToNextPage, changeNumberOfRows, firstIndex, secondIndex} = props; 
     
     return (
