+++ src/pages/students/students-list/student-table/student-table.scss
@@ -1,4 +1,4 @@
-@use "../../../assets/styles/_typography" as *; 
+@use "../../../../assets/styles/_typography" as *; 
 
 a{
   text-decoration: none;
