+++ src/pages/students/students-list/students.scss
@@ -1,4 +1,4 @@
-@import "../../assets/styles/variables.scss";
+@import "../../../assets/styles/variables";
 .students {
   width: 100%; 
   background-color: $color-primary;
