+++ src/pages/students/students-list/studentsearchbar/studentsearchbar.jsx
@@ -1,8 +1,9 @@
 import React from 'react'
-import Filter from "../../../assets/images/filter.png";
-import GridView from "../../../assets/images/grid-view.png";
-import HamburgerMenu from "../../../assets/images/hamburger-menu.png";
-import Sort from "../../../assets/images/sort.png";
+
+import Filter from "../../../../assets/images/filter.png";
+import GridView from "../../../../assets/images/grid-view.png";
+import HamburgerMenu from "../../../../assets/images/hamburger-menu.png";
+import Sort from "../../../../assets/images/sort.png";
 import "./studentsearchbar.scss";
