+++ src/pages/students/students-list/studentsearchbar/studentsearchbar.scss
@@ -1,4 +1,4 @@
-@import "../../../assets/styles/_variables";
+@import "../../../../assets/styles/_variables";
 
 .search-bar {
   width: 100%; 
@@ -11,7 +11,7 @@
     }
   }
   &__input {
-    background-image: url("../../../assets/images/search.png");
+    background-image: url("../../../../assets/images/search.png");
     background-repeat: no-repeat;
     background-position: 97% center;
     border: none;
