+++ src/pages/students/students-list/studentstopnav/studentstopnav.scss
@@ -1,5 +1,5 @@
-@use "../../../assets/styles/buttons";
-@use "../../../assets/styles/_variables.scss" as *;
+@use "../../../../assets/styles/buttons";
+@use "../../../../assets/styles/_variables.scss" as *;
 
 .top-nav{
     width: 100%;
@@ -7,6 +7,7 @@
     padding-top: 59px;
     background:#FFFFFF;
     margin-bottom: 30px; 
+    
     &__header {
         width: 100%; 
         height: 50px;
